requirements.txt
aniso8601==9.0.1
bcrypt==4.2.0
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
Flask==3.0.2
Flask-Cors==4.0.0
Flask-RESTful==0.3.10
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
mysql-connector-python==8.1.0
protobuf==4.21.12
PyJWT==2.9.0
python-dotenv==1.0.1
pytz==2024.1
requests==2.31.0
six==1.16.0
urllib3==2.2.1
Werkzeug==3.0.1
